requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
degiro-connector>=2.0.0
yfinance>=0.2.30
python-dotenv>=1.0.0
